+++ src/apps/replication/lib/mutation.h
@@ -98,7 +98,7 @@ inline void mutation::set_id(ballot b, decree c)
 {
     data.header.ballot = b;
     data.header.decree = c;
-    sprintf (_name, "%lld.%lld", b, c);
+    sprintf (_name, "%lld.%lld", static_cast<long long int>(b), static_cast<long long int>(c));
 }
 
 }} // namespace
